adapted backend to make working the welcome page and refactor server
