+++ package.json
@@ -7,5 +7,5 @@
     "type": "git",
     "url": "https://github.com/snipsco/teleport-flask-webrouter"
   },
-	"version": "0.1.8"
+	"version": "0.1.9"
 }
